new organization for URLConnectionTest test case
